requirements.txt
kivy
kivymd
pyyaml
python-dotenv
flask
pyjnius
cython
buildozer
pytest
cryptography
aiortc
qrcode
pyzbar
pillow
firebase-admin
pydantic
google-auth 
google-auth-oauthlib
google-auth-httplib2
pytest
bcrypt
keyring
keyrings.alt